backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Azure OpenAI
openai==1.3.0
azure-identity==1.15.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Database & RAG
qdrant-client==1.15.1
onnxruntime==1.16.3
onnx==1.15.0
langchain==0.1.0
pypdf==3.17.1
huggingface-hub==0.16.4
transformers==4.35.0
sentencepiece==0.1.99
rank-bm25==0.2.2
scikit-learn==1.3.2
networkx==3.2.1  # Graph RAG knowledge graph

# Document Processing
python-docx==1.1.0
beautifulsoup4==4.12.2
markdownify==0.11.6
openpyxl==3.1.2

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
tenacity==8.2.3
numpy>=1.24.0
tiktoken==0.12.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# OpenTelemetry (Distributed Tracing)
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-prometheus==0.43b0

# Data Quality Monitoring
evidently==0.4.15

# RAG Evaluation
ragas==0.1.4
datasets==2.16.1

# SSE Support
sse-starlette==1.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
